cmd: Add cilium node neigh CLI
This allows users to manage the current node's neighbor table (the node
that's receiving this request, i.e. the node running Cilium) from the
CLI. The only supported operations at the moment are inserting and
removing a node into / from the current node's neighbor table. See
previous commit for context into why we are exposing this API.

Signed-off-by: Chris Tarazi <chris@isovalent.com>